refactor(editor): move viewport renderer to gfx
