+++ blocksuite/affine/shared/package.json
@@ -63,7 +63,6 @@
     "./theme": "./src/theme/index.ts",
     "./styles": "./src/styles/index.ts",
     "./services": "./src/services/index.ts",
-    "./viewport-renderer": "./src/viewport-renderer/index.ts",
     "./adapters": "./src/adapters/index.ts"
   },
   "files": [
