--- blocksuite/affine/shared/src/viewport-renderer/index.ts
@@ -1,2 +1,0 @@
-export * from './types.js';
-export * from './viewport-renderer.js';
