+++ blocksuite/framework/block-std/src/gfx/viewport/turbo-renderer/text-utils.ts
